Updated interpreter with benchmarking option
Can be triggered by inserting the ":benchmark" key inside the starting state of the push VM which leaves the measurements in the final VM state under the key "efficiency" as a hashmap of ":runtime" and ":memory". This works very similarly to the already existing ":keep-history" parameter in the interpreter. The handling of multiple parameters is not too elegant, but can be improved later.
Importantly, I haven't tested whether this works properly, so I'm creating a new branch for it and it should only be merged when we verify that it works correctly. Note: this uses criterium to benchmark runtimes, which will yield very precise results that account for any JVM bottlenecks. On the other hand, getting a precise runtime takes long, so we should potentially talk about other ways to do it.